move enum to correct folder
