+++ models/wql/__init__.py
@@ -8,13 +8,13 @@ from models.wql.clauses.make.make_pre import (
 from models.wql.clauses.make.make_properties import (
     MakeProperty,
     ExtractedProperty,
-    PropertyType,
 )
 from models.wql.data import WG_ALLOWED_TYPES
 from models.wql.data.node import Node
 from models.wql.data.relationship import Relationship
 from models.wql.data.wiggle_metadata import WiggleGraphMetalData
 from models.wql.enums.clauses import Clause
+from models.wql.enums.property_type import PropertyType
 from models.wql.parsed_query import ParsedMake, ParsedPattern, ParsedQuery
 
 __all__ = [
