+++ models/wql/enums/property_type.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class PropertyType(Enum):
+    NONE_TYPE = "NONE_TYPE"
+    BOOL_TYPE = "BOOL_TYPE"
+    FLOAT_TYPE = "FLOAT_TYPE"
+    INT_TYPE = "INT_TYPE"
+    LIST_TYPE = "LIST_TYPE"
+    STRING_TYPE = "STRING_TYPE"
